requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
pymongo
langchain
langchain-community
langchain-openai
groq
sqlalchemy
